# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Kanban API is an restful API for task management for developers or any corperate workers\n\nIt is a standalone API",
    "title": "KanBan API",
    "contact": {
      "name": "John Ayilara",
      "url": "https://bolu.netlify.app",
      "email": "jbayilara@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost",
  "basePath": "/api/v1/",
  "paths": {
...
    "/tasks": {
      "get": {
        "description": "Get all the tasks for the user",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "This route get all the tasks created by the user",
        "operationId": "GetAllTasks",
        "parameters": [
          {
            "type": "string",
            "description": "\"Authorization token\"",
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "description": "\"Authorization token\"",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskReply"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/tasks/:id": {
      "get": {
        "description": "Get a task",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "This route get a task",
        "operationId": "GetTask",
        "parameters": [
          {
            "type": "string",
            "description": "\"Authorization token\"",
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "description": "\"Authorization token\"",
              "type": "string"
            }
          },
          {
            "type": "string",
            "description": "\"Task ID\"",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "description": "\"Task ID\"",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskReply"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "put": {
        "description": "Update a task",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "This route update a task",
        "operationId": "UpdateTask",
        "parameters": [
          {
            "type": "string",
            "description": "\"Authorization token\"",
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "description": "\"Authorization token\"",
              "type": "string"
            }
          },
          {
            "type": "string",
            "description": "\"Task ID\"",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "description": "\"Task ID\"",
              "type": "string"
            }
          },
          {
            "type": "object",
            "description": "\"Task\"",
            "name": "task",
            "in": "body",
            "required": true,
            "schema": {
              "description": "\"Task\"",
              "type": "object",
              "$ref": "#/definitions/TaskRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskReply"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "description": "Create a task",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "This route create a task",
        "operationId": "CreateTask",
        "parameters": [
          {
            "type": "string",
            "description": "\"Authorization token\"",
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "description": "\"Authorization token\"",
              "type": "string"
            }
          },
          {
            "type": "string",
            "description": "\"Board ID\"",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "description": "\"Board ID\"",
              "type": "string"
            }
          },
          {
            "type": "object",
            "description": "\"Task to be created\"",
            "name": "task",
            "in": "body",
            "required": true,
            "schema": {
              "description": "\"Task to be created\"",
              "type": "object",
              "$ref": "#/definitions/TaskRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskReply"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      },
      "delete": {
        "description": "Delete a task",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "This route delete a task",
        "operationId": "DeleteTask",
        "parameters": [
          {
            "type": "string",
            "description": "\"Authorization token\"",
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "description": "\"Authorization token\"",
              "type": "string"
            }
          },
          {
            "type": "string",
            "description": "\"Task ID\"",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "description": "\"Task ID\"",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskReply"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
    "/tasks/:status": {
      "delete": {
        "description": "Delete the task with the specified status",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tasks"
        ],
        "summary": "This route delete task with a particular status",
        "operationId": "DeleteTask",
        "parameters": [
          {
            "type": "string",
            "description": "\"Authorization token\"",
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "description": "\"Authorization token\"",
              "type": "string"
            }
          },
          {
            "type": "string",
            "description": "\"Task status\"",
            "name": "status",
            "in": "path",
            "required": true,
            "schema": {
              "description": "\"Task status\"",
              "type": "string"
            }
          },
          {
            "type": "string",
            "description": "\"User Id\"",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "description": "\"User Id\"",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/TaskReply"
          },
          "500": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    }
  },
  "definitions": {
...
    "ObjectID": {
      "type": "array",
      "title": "ObjectID is the BSON ObjectID type.",
      "items": {
        "type": "integer",
        "format": "uint8"
      },
      "x-go-package": "go.mongodb.org/mongo-driver/bson/primitive"
    },
    "SubTask": {
      "type": "object",
      "properties": {
        "id": {
          "$ref": "#/definitions/ObjectID"
        },
        "is_done": {
          "type": "boolean",
          "x-go-name": "IsDone"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        }
      },
      "x-go-package": "kanban-task/model"
    },
    "Task": {
      "type": "object",
      "properties": {
        "board_id": {
          "$ref": "#/definitions/ObjectID"
        },
        "description": {
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "$ref": "#/definitions/ObjectID"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        },
        "sub_tasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SubTask"
          },
          "x-go-name": "SubTasks"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        },
        "user_id": {
          "type": "string",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "kanban-task/model"
    },
    "TaskRequest": {
      "type": "object",
      "required": [
        "title",
        "description",
        "status",
        "sub_tasks"
      ],
      "properties": {
        "description": {
          "description": "The description for the task",
          "type": "string",
          "x-go-name": "Description",
          "example": "My Task Description"
        },
        "status": {
          "description": "The status for the task",
          "type": "string",
          "x-go-name": "Status",
          "example": "TODO"
        },
        "sub_tasks": {
          "description": "The sub tasks for the task",
          "type": "array",
          "items": {
            "$ref": "#/definitions/SubTask"
          },
          "x-go-name": "SubTasks",
          "example": []
        },
        "title": {
          "description": "The title for the task",
          "type": "string",
          "x-go-name": "Title",
          "example": "My Task"
        }
      },
      "x-go-package": "kanban-task/model"
    }
  },
  "responses": {
...
    "ErrorResponse": {
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "description": "The error message",
            "type": "string",
            "x-go-name": "Message",
            "example": "Error message"
          }
        }
      }
    },
    "TaskReply": {
      "schema": {
        "$ref": "#/definitions/Task"
      }
    },
...
  }
}